Update bottom sheet with rep/weight

+++ app/src/main/java/com/autilite/weightlifttracker/program/Exercise.java
@@ -59,7 +59,7 @@ public class Exercise {
         return weight;
     }
 
-    public void setWeight(int weight) {
+    public void setWeight(float weight) {
         this.weight = weight;
     }
